added login methods and updated old test files
