+++ spec/selenium/moveMarkers_test.rb
@@ -5,9 +5,8 @@ require "selenium-webdriver"
 require "rspec"
 require "chromedriver-helper"
 
-# auth id and pass
-$login_id = 'msprayapptest@gmail.com'
-$login_pass = 'mSprayApp2.0'
+# log in methods
+require_relative "log_in.rb"
 
 # move offset values
 $x_offset = 100
@@ -15,7 +14,7 @@ $y_offset = 100
 
 describe "automating mSpray QA" do
   it "moves markers" do
-    setup
+    @driver = set_up
     move_marker
     expect(@driver.find_element(css: '.fas.fa-map').displayed?)
     @driver.quit
@@ -39,58 +38,4 @@ def move_marker
   @driver.action.send_keys(modal, :tab).perform
   @driver.action.send_keys(:enter).perform
   waitUntil(30, 'map')
-end
-
-def setup
-  # sets up selenium webdriver test environment
-  @driver = Selenium::WebDriver.for :chrome
-  hack_Google
-  dismiss_notification
-end
-
-def waitUntil(time, element)
-  # waits until a predefined element displayed
-  wait = Selenium::WebDriver::Wait.new(timeout: time)
-
-  # Google login id
-  if element == 'login'
-    wait.until { @driver.find_element(id: 'identifierId').displayed? }
-  # Google login password
-  elsif element == 'password'
-    wait.until { @driver.find_element(id: 'password').displayed? }
-  # mSpray map element
-  elsif element == 'map'
-    wait.until { @driver.find_element(css: '.fas.fa-map').displayed? }
-  # flags on map
-  elsif element == 'button'
-    wait.until { @driver.find_element(css: '[role="button"]').displayed? }
-  # delete marker modal
-  elsif element == 'modal'
-    wait.until { @driver.find_element(id: 'move_pin').displayed? }
-  end
-end
-
-def hack_Google
-  # navigates to main page
-  @driver.get "http://localhost:3000/"
-
-  # logs into Google Accounts
-  # inputs id and waits for password prompt
-  @driver.find_element(css: '.btn-large.red').click
-  waitUntil(30, 'login')
-  @driver.find_element(id: 'identifierId').send_keys($login_id)
-  @driver.find_element(id: 'identifierNext').click
-  waitUntil(10, 'password')
-
-  # inputs password and waits for map display
-  @driver.find_element(name: 'password').send_keys($login_pass)
-  @driver.find_element(id: 'passwordNext').click
-  waitUntil(30, 'map')
-end
-
-def dismiss_notification
-  # dismisses error messages
-  if @driver.find_element(css: '.dismissButton').displayed?
-    @driver.find_element(css: '.dismissButton').click
-  end
 end
